refactor(logic): 피드백 반영
Pattern 대신 Regex를 활용한다. 단 Regex를 매번 컴파일 하지 않도록 미리 만들어둔다.
StringRes는 View 로직이므로 옮긴다.
