+++ src/main/kotlin/calculator/parser/NodeParser.kt
@@ -3,24 +3,22 @@ package calculator.parser
 import calculator.calculator.Node
 import calculator.calculator.Number
 import calculator.calculator.Operator
-import java.util.regex.Pattern
 
 class NodeParser {
     fun parse(text: String): List<Node> {
-        return MATCHER.split(text)
+        return REGEX_FOR_NODE.split(text)
             .trimAll()
             .exceptEmpty()
             .convertAllToNode()
     }
 
-    private fun Array<String>.trimAll() = map { it.trim() }
+    private fun List<String>.trimAll() = map { it.trim() }
 
     private fun List<String>.exceptEmpty() = filter { it.isNotEmpty() }
 
     private fun String.convertToNode(): Node {
         takeIf {
-            Pattern.compile(PATTERN_FOR_NUMBER)
-                .matcher(it).matches()
+            REGEX_FOR_NUMBER.matches(it)
         }?.let {
             return Number(it)
         }
@@ -31,8 +29,9 @@ class NodeParser {
     private fun List<String>.convertAllToNode() = map { it.convertToNode() }
 
     companion object {
-        private const val REGEX_FOR_NODE = "((?<=([-*/+\\s]))|(?=([-*/+\\s])))"
-        private val MATCHER = Pattern.compile(REGEX_FOR_NODE)
+        private const val REGEX_STRING = "((?<=([-*/+\\s]))|(?=([-*/+\\s])))"
+        private val REGEX_FOR_NODE = Regex(REGEX_STRING)
         private const val PATTERN_FOR_NUMBER = "^[0-9]+$"
+        private val REGEX_FOR_NUMBER = Regex(PATTERN_FOR_NUMBER)
     }
 }
