+++ src/main/kotlin/racingcar/view/ResultView.kt
@@ -1,8 +1,6 @@
 package racingcar.view
 
 import racingcar.domain.Car
-import racingcar.domain.WINNER_IS
-import racingcar.domain.WINNER_SEPARATOR
 
 class ResultView {
     fun showResult(racingResult: String, winnerResult: List<Car>) {
